A100_submissions/A100_2nd_submission/requirements.txt
torch
protobuf
sentencepiece
ctranslate2
transformers
ipdb
gdown
bert_score
faiss-gpu
datasets
tiktoken
einops 
ninja
packaging
transformers_stream_generator
nltk
